fix: some circular deps
