+++ src/cli-ux/styled/json.ts
@@ -1,15 +1,17 @@
 import chalk from 'chalk'
+import {format} from 'node:util'
+import {stdout} from '../stream'
 
-import {ux} from '../../index'
+const info = (output: string) => stdout.write(format(output) + '\n')
 
 export default function styledJSON(obj: unknown): void {
   const json = JSON.stringify(obj, null, 2)
   if (!chalk.level) {
-    ux.info(json)
+    info(json)
     return
   }
 
   const cardinal = require('cardinal')
   const theme = require('cardinal/themes/jq')
-  ux.info(cardinal.highlight(json, {json: true, theme}))
+  info(cardinal.highlight(json, {json: true, theme}))
 }
