+++ src/config/index.ts
@@ -1,5 +1,4 @@
 export {Config} from './config'
-export {toCached} from './config'
 export {Plugin} from './plugin'
 export {tsPath} from './ts-node'
