+++ src/errors/handle.ts
@@ -2,7 +2,7 @@
 /* eslint-disable unicorn/no-process-exit */
 import {OclifError, PrettyPrintableError} from '../interfaces'
 import {CLIError} from './errors/cli'
-import {ExitError} from '.'
+import {ExitError} from './errors/exit'
 import clean from 'clean-stack'
 import {config} from './config'
 import prettyPrint from './errors/pretty-print'
