+++ src/index.ts
@@ -19,7 +19,8 @@ export * as Flags from './flags'
 export * as Parser from './parser'
 export * as ux from './cli-ux'
 export {CommandHelp, HelpBase, Help, loadHelpClass} from './help'
-export {Config, toCached, Plugin, tsPath} from './config'
+export {Config, Plugin, tsPath} from './config'
+export {toCached} from './to-cached'
 export {HelpSection, HelpSectionRenderer, HelpSectionKeyValueTable} from './help/formatter'
 export {Settings, settings} from './settings'
 export {stdout, stderr} from './cli-ux/stream'
